Temporary Fix for i-RIC/iricdev#5
uses https://github.com/scharlton2/CGNS/tree/v3.2.1-patch1 to
replace calls to stat with _stat64

+++ build-cgnslib.cmd
@@ -9,12 +9,24 @@ if [%GENERATOR%]==[] (
 call versions.cmd
 set VER=%CGNSLIB_VER%
 
-if exist lib\src\cgnslib-%VER%     rd /s /q lib\src\cgnslib-%VER%
+if "%CGNSLIB_VER%"=="3.2.1" (
+  set VER=%CGNSLIB_VER%-patch1
+  if exist lib\src\cgnslib-%VER%     echo WARNING lib\src\cgnslib-%VER% exists NOT DELETING!
+) else (
+  if exist lib\src\cgnslib-%VER%     rd /s /q lib\src\cgnslib-%VER%
+)
+
 if exist lib\build\cgnslib-%VER%   rd /s /q lib\build\cgnslib-%VER%
 if exist lib\install\cgnslib-%VER% rd /s /q lib\install\cgnslib-%VER%
 
-7z x cgnslib_%VER%.tar.gz -so | 7z x -si -ttar -o"lib\src"
-ren lib\src\cgnslib_%VER% cgnslib-%VER%
+if "%CGNSLIB_VER%"=="3.2.1" (
+  if NOT exist lib\src\cgnslib-3.2.1-patch1 (
+    git clone --branch v3.2.1-patch1 git@github.com:scharlton2/CGNS.git lib/src/cgnslib-%VER%
+  )
+) else (
+  7z x cgnslib_%VER%.tar.gz -so | 7z x -si -ttar -o"lib\src"
+  ren lib\src\cgnslib_%VER% cgnslib-%VER%
+)
 
 ctest -S build-cgnslib.cmake -DCONF_DIR:STRING=debug   -DCTEST_CMAKE_GENERATOR:STRING=%GENERATOR% -C Debug   -VV -O %SGEN%-cgnslib-debug.log
 ctest -S build-cgnslib.cmake -DCONF_DIR:STRING=release -DCTEST_CMAKE_GENERATOR:STRING=%GENERATOR% -C Release -VV -O %SGEN%-cgnslib-release.log
